解决详情页的空指针异常的bug与隐患

+++ cx-bbs-model/src/main/java/com/cx/model/article/pojos/Question.java
@@ -38,7 +38,7 @@ public class Question implements Serializable {
     private String title;
 
     @ApiModelProperty("文章作者的ID")
-    private Long authorId;
+    private Integer authorId;
 
     @ApiModelProperty("点赞数量")
     private Integer likes;
